Convert all remaining JS tooling files to TypeScript; remove Babel
All JavaScript files in the client package are now TypeScript:
- vite.config.js → vite.config.ts
- vitest.config.js → vitest.config.ts
- prettier.config.mjs → prettier.config.ts (adds Config type annotation)
- eslint.config.mjs → eslint.config.ts (removes Babel parser; now uses
  @typescript-eslint/parser for both .ts and .vue files since all SFCs
  have <script lang="ts">; requires jiti for ESLint TypeScript config loading)
- scripts/generate-doc-manifest.js → generate-doc-manifest.ts (converts
  CommonJS require() to ESM imports; adds types; runs via node --strip-types)
- babel.config.js → deleted

Package changes:
- Remove @babel/core, @babel/eslint-parser, @babel/preset-env
- Add jiti (required by ESLint 9.39+ for TypeScript config file loading)
- Add "type": "module" to eliminate Node ESM detection warnings
- Update lint/format globs to drop .js (no .js files remain)
- Update prebuild script to use node --strip-types for the .ts manifest script

tsconfig.json: remove allowJs / checkJs (no .js source files remain)
