--- client/babel.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  presets: [
-    '@babel/preset-env',
-  ],
-};
