+++ client/package.json
@@ -4,17 +4,18 @@
   "description": "DigiScript front end",
   "author": "DreamTeamProd",
   "private": true,
+  "type": "module",
   "scripts": {
-    "prebuild": "scripts/copy-docs.sh && node scripts/generate-doc-manifest.js",
+    "prebuild": "scripts/copy-docs.sh && node --strip-types scripts/generate-doc-manifest.ts",
     "build": "vite build",
     "build:analyze": "vite build --mode analyze",
     "lint": "npm run format && npm run lint:eslint",
-    "lint:eslint": "eslint 'src/**/*.{js,ts,vue}' --fix",
+    "lint:eslint": "eslint 'src/**/*.{ts,vue}' --fix",
     "ci-lint": "npm run format:check && npm run lint:eslint-check",
-    "lint:eslint-check": "eslint 'src/**/*.{js,ts,vue}'",
+    "lint:eslint-check": "eslint 'src/**/*.{ts,vue}'",
     "lint:filter": "./scripts/eslint-filter.sh",
-    "format": "prettier --write 'src/**/*.{js,ts,vue,json,css,scss}'",
-    "format:check": "prettier --check 'src/**/*.{js,ts,vue,json,css,scss}'",
+    "format": "prettier --write 'src/**/*.{ts,vue,json,css,scss}'",
+    "format:check": "prettier --check 'src/**/*.{ts,vue,json,css,scss}'",
     "typecheck": "tsc --noEmit -p tsconfig.json",
     "test": "vitest",
     "test:ui": "vitest --ui",
@@ -53,9 +54,6 @@
     "vuex-persistedstate": "3.2.1"
   },
   "devDependencies": {
-    "@babel/core": "7.29.0",
-    "@babel/eslint-parser": "7.28.6",
-    "@babel/preset-env": "7.29.5",
     "@eslint/js": "^9.39.2",
     "@types/lodash": "~4.17.7",
     "@types/node": "~20.14.0",
@@ -70,6 +68,7 @@
     "eslint-plugin-prettier": "^5.5.5",
     "eslint-plugin-vue": "^9.33.0",
     "globals": "^15.14.0",
+    "jiti": "^2.7.0",
     "jsdom": "^27.4.0",
     "prettier": "^3.8.3",
     "sass": "1.99.0",
