+++ client/tsconfig.json
@@ -29,8 +29,6 @@
     // skipLibCheck: third-party packages (bootstrap-vue, vue-native-websocket,
     // @types/vuelidate) carry incomplete or conflicting .d.ts files.
     "skipLibCheck": true,
-    "allowJs": true,      // Retained for vite.config.js and other tooling config files outside src/
-    "checkJs": false,     // Don't type-check .js files
     "jsx": "preserve",    // Vue template compiler handles JSX transform
     "baseUrl": ".",
     "paths": {
