+++ client/vite.config.ts
